+++ changelog.md
@@ -1,6 +1,14 @@
 ## Changes in version 0.6.4.1
 
  * Add `Prim` instance for `Complex`.
+ 
+ * Add instances for the following newtypes from `base`:
+   `Const`, `Identity`, `Down`, `Dual`, `Sum`, `Product`,
+   `First`, `Last`, `Min`, `Max`
+
+ * Add `base-orphans` dependency to test suite to accomodate
+   older versions of GHC not having instances of `Show` and `Eq`
+   for some of the above newtypes.
 
 ## Changes in version 0.6.4.0