+++ changelog.md
@@ -1,6 +1,8 @@
 ## Changes in version 0.6.4.1
 
- * Add instances for the following newtypes from `base`:
+ * Add `foldMapPrimArray` and `foldMapPrimArray'`
+
+ * Add `Prim` instances for the following newtypes from `base`:
    `Const`, `Identity`, `Down`, `Dual`, `Sum`, `Product`,
    `First`, `Last`, `Min`, `Max`